2025 Nov 21 (Fri)

+++ desempenho/simulador/trabalho2/ExclamacaoAtena/src/measurement_window.cpp
@@ -1,11 +1,23 @@
+/**
+ * @file measurement_window.cpp
+ * @brief Implementation of circular measurement window
+ */
+
 #include "../include/measurement_window.hpp"
 #include <stdexcept>
 
+/**
+ * @brief Construct a new Circular Measurement Window object
+ */
 CircularMeasurementWindow::CircularMeasurementWindow()
     : headIndex(0), tailIndex(0), currentSize(0), maxSize(0),
       sumArrivalTimestamps(0.0), sumWaitingTimes(0.0), totalPacketsInWindow(0) {
 }
 
+/**
+ * @brief Initialize measurement window
+ * @param size Maximum window size
+ */
 void CircularMeasurementWindow::initialize(int size) {
     maxSize = size;
     arrivalTimestamps.resize(maxSize, 0.0);
@@ -20,6 +32,12 @@ void CircularMeasurementWindow::initialize(int size) {
     totalPacketsInWindow = 0;
 }
 
+/**
+ * @brief Add packet metadata to window
+ * @param arrivalTime Packet arrival timestamp
+ * @param departureTime Packet departure timestamp
+ * @param waitingTime Calculated waiting time
+ */
 void CircularMeasurementWindow::addPacket(double arrivalTime, double departureTime, double waitingTime) {
     // If buffer is full, remove oldest element
     if (currentSize == maxSize) {
@@ -44,16 +62,27 @@ void CircularMeasurementWindow::addPacket(double arrivalTime, double departureTi
     currentSize++;
 }
 
+/**
+ * @brief Compute average waiting time in window
+ * @return double Average waiting time
+ */
 double CircularMeasurementWindow::computeWindowAverageWaitingTime() const {
     if (totalPacketsInWindow == 0) return 0.0;
-    return sumWaitingTimes / totalPacketsInWindow;
+    return sumWaitingTimes / static_cast<double>(totalPacketsInWindow);
 }
 
+/**
+ * @brief Compute average arrival time in window
+ * @return double Average arrival time
+ */
 double CircularMeasurementWindow::computeWindowAverageArrivalTime() const {
     if (totalPacketsInWindow == 0) return 0.0;
-    return sumArrivalTimestamps / totalPacketsInWindow;
+    return sumArrivalTimestamps / static_cast<double>(totalPacketsInWindow);
 }
 
+/**
+ * @brief Clear all window data
+ */
 void CircularMeasurementWindow::clear() {
     initialize(maxSize);
 }
